+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Main
+
+##### Breaking
+- The `dump` functions now requires `NodeRepresentable` values rather than `Any?` (or any `Sequence`) [#299]
+
 ## 4.0.6
 
 ##### Breaking